Toy program with Eigen::Map on a SGMatrix.
During the Shogun hackaton in Zurich. We wanted to see what happens
with the underlying data in SGMatrix when Eigen::Map is applied to it.
The conclusion is that the SGMatrix is not touched (at least when
using transpose) so it must be that a copy is made when calling tranpose.